requirements.txt
langchain
langgraph
langchain-core
langchain-community
python-dotenv
langchain-groq
pandas
numpy 
scikit-learn
streamlit